moved helper functions to different file
